💄 style(input): update background-color logic for block type input
This commit updates the background-color logic for block type input to use `token.colorFillTertiary` instead of `rgba(token.colorBgElevated, 0.6)` in `src/Highlighter/style.ts` and `src/Input/style.ts`. The change is made in order to improve the readability of the code.
